LayoutTests/css3/flexbox/css-properties-expected.txt

Tests being able to set the display to -webkit-flex and -webkit-inline-flex.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS flexbox.style.display is "-webkit-flex"
PASS flexbox.style.display is "inline"
PASS flexbox.style.display is "-webkit-inline-flex"
PASS flexbox.style.display is "-webkit-inline-flex"
PASS flexbox.style.display is "block"
PASS flexitem.style.webkitOrder is ""
PASS flexitem.style.webkitOrder is "2"
PASS flexitem.style.webkitOrder is "-1"
PASS flexitem.style.webkitOrder is "0"
PASS flexitem.style.webkitOrder is "0"
PASS flexitem.style.webkitOrder is "0"
PASS flexitem.style.webkitOrder is ""
PASS flexbox.style.webkitFlexPack is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexPack is "start"
PASS flexbox.style.webkitFlexPack is ""
PASS flexbox.style.webkitFlexPack is "start"
PASS window.getComputedStyle(flexbox, null).webkitFlexPack is "start"
PASS flexbox.style.webkitFlexPack is "end"
PASS window.getComputedStyle(flexbox, null).webkitFlexPack is "end"
PASS flexbox.style.webkitFlexPack is "center"
PASS window.getComputedStyle(flexbox, null).webkitFlexPack is "center"
PASS flexbox.style.webkitFlexPack is "space-between"
PASS window.getComputedStyle(flexbox, null).webkitFlexPack is "space-between"
PASS flexbox.style.webkitFlexPack is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexPack is "start"
PASS flexbox.style.webkitAlignSelf is ""
PASS window.getComputedStyle(flexbox, null).webkitAlignSelf is "stretch"
PASS window.getComputedStyle(document.documentElement, null).webkitAlignSelf is "stretch"
PASS flexbox.style.webkitAlignSelf is ""
PASS flexbox.style.webkitAlignSelf is "auto"
PASS window.getComputedStyle(flexbox, null).webkitAlignSelf is "stretch"
PASS flexbox.style.webkitAlignSelf is "start"
PASS window.getComputedStyle(flexbox, null).webkitAlignSelf is "start"
PASS flexbox.style.webkitAlignSelf is "end"
PASS window.getComputedStyle(flexbox, null).webkitAlignSelf is "end"
PASS flexbox.style.webkitAlignSelf is "center"
PASS window.getComputedStyle(flexbox, null).webkitAlignSelf is "center"
PASS flexbox.style.webkitAlignSelf is "stretch"
PASS window.getComputedStyle(flexbox, null).webkitAlignSelf is "stretch"
PASS flexbox.style.webkitAlignSelf is "baseline"
PASS window.getComputedStyle(flexbox, null).webkitAlignSelf is "baseline"
PASS flexbox.style.webkitAlignSelf is ""
PASS window.getComputedStyle(flexbox, null).webkitAlignSelf is "stretch"
PASS flexbox.style.webkitAlignItems is ""
PASS window.getComputedStyle(flexbox, null).webkitAlignItems is "stretch"
PASS window.getComputedStyle(flexitem, null).webkitAlignSelf is "stretch"
PASS flexbox.style.webkitAlignItems is ""
PASS flexbox.style.webkitAlignItems is ""
PASS window.getComputedStyle(flexbox, null).webkitAlignItems is "stretch"
PASS window.getComputedStyle(flexitem, null).webkitAlignSelf is "stretch"
PASS flexbox.style.webkitAlignItems is "start"
PASS window.getComputedStyle(flexbox, null).webkitAlignItems is "start"
PASS window.getComputedStyle(flexitem, null).webkitAlignSelf is "start"
PASS flexbox.style.webkitAlignItems is "end"
PASS window.getComputedStyle(flexbox, null).webkitAlignItems is "end"
PASS window.getComputedStyle(flexitem, null).webkitAlignSelf is "end"
PASS flexbox.style.webkitAlignItems is "center"
PASS window.getComputedStyle(flexbox, null).webkitAlignItems is "center"
PASS window.getComputedStyle(flexitem, null).webkitAlignSelf is "center"
PASS flexbox.style.webkitAlignItems is "stretch"
PASS window.getComputedStyle(flexbox, null).webkitAlignItems is "stretch"
PASS window.getComputedStyle(flexitem, null).webkitAlignSelf is "stretch"
PASS flexbox.style.webkitAlignItems is "baseline"
PASS window.getComputedStyle(flexbox, null).webkitAlignItems is "baseline"
PASS window.getComputedStyle(flexitem, null).webkitAlignSelf is "baseline"
PASS flexbox.style.webkitAlignItems is ""
PASS window.getComputedStyle(flexbox, null).webkitAlignItems is "stretch"
PASS window.getComputedStyle(flexitem, null).webkitAlignSelf is "stretch"
PASS flexbox.style.webkitAlignItems is ""
PASS window.getComputedStyle(flexbox, null).webkitAlignItems is "stretch"
PASS window.getComputedStyle(flexitem, null).webkitAlignSelf is "stretch"
PASS window.getComputedStyle(detachedFlexbox, null).webkitAlignSelf is ""
PASS window.getComputedStyle(detachedFlexItem, null).webkitAlignSelf is ""
PASS flexbox.style.webkitFlexDirection is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexDirection is "row"
PASS flexbox.style.webkitFlexDirection is ""
PASS flexbox.style.webkitFlexDirection is "row"
PASS window.getComputedStyle(flexbox, null).webkitFlexDirection is "row"
PASS flexbox.style.webkitFlexDirection is "row-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexDirection is "row-reverse"
PASS flexbox.style.webkitFlexDirection is "column"
PASS window.getComputedStyle(flexbox, null).webkitFlexDirection is "column"
PASS flexbox.style.webkitFlexDirection is "column-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexDirection is "column-reverse"
PASS flexbox.style.webkitFlexWrap is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexWrap is "none"
PASS flexbox.style.webkitFlexWrap is ""
PASS flexbox.style.webkitFlexWrap is "none"
PASS window.getComputedStyle(flexbox, null).webkitFlexWrap is "none"
PASS flexbox.style.webkitFlexWrap is "wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexWrap is "wrap"
PASS flexbox.style.webkitFlexWrap is "wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexWrap is "wrap-reverse"
PASS flexbox.style.webkitFlexFlow is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row none"
PASS flexbox.style.webkitFlexFlow is ""
PASS flexbox.style.webkitFlexFlow is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row none"
PASS flexbox.style.webkitFlexFlow is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row none"
PASS flexbox.style.webkitFlexFlow is "none"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row none"
PASS flexbox.style.webkitFlexFlow is "none"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row none"
PASS flexbox.style.webkitFlexFlow is "wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row wrap"
PASS flexbox.style.webkitFlexFlow is "wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row wrap"
PASS flexbox.style.webkitFlexFlow is "wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row wrap-reverse"
PASS flexbox.style.webkitFlexFlow is "wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row wrap-reverse"
PASS flexbox.style.webkitFlexFlow is "row"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row none"
PASS flexbox.style.webkitFlexFlow is "row"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row none"
PASS flexbox.style.webkitFlexFlow is "row none"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row none"
PASS flexbox.style.webkitFlexFlow is "row none"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row none"
PASS flexbox.style.webkitFlexFlow is "row wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row wrap"
PASS flexbox.style.webkitFlexFlow is "row wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row wrap"
PASS flexbox.style.webkitFlexFlow is "row wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row wrap-reverse"
PASS flexbox.style.webkitFlexFlow is "row wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row wrap-reverse"
PASS flexbox.style.webkitFlexFlow is "row-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row-reverse none"
PASS flexbox.style.webkitFlexFlow is "row-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row-reverse none"
PASS flexbox.style.webkitFlexFlow is "row-reverse none"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row-reverse none"
PASS flexbox.style.webkitFlexFlow is "row-reverse none"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row-reverse none"
PASS flexbox.style.webkitFlexFlow is "row-reverse wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row-reverse wrap"
PASS flexbox.style.webkitFlexFlow is "row-reverse wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row-reverse wrap"
PASS flexbox.style.webkitFlexFlow is "row-reverse wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row-reverse wrap-reverse"
PASS flexbox.style.webkitFlexFlow is "row-reverse wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row-reverse wrap-reverse"
PASS flexbox.style.webkitFlexFlow is "column"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column none"
PASS flexbox.style.webkitFlexFlow is "column"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column none"
PASS flexbox.style.webkitFlexFlow is "column none"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column none"
PASS flexbox.style.webkitFlexFlow is "column none"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column none"
PASS flexbox.style.webkitFlexFlow is "column wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column wrap"
PASS flexbox.style.webkitFlexFlow is "column wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column wrap"
PASS flexbox.style.webkitFlexFlow is "column wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column wrap-reverse"
PASS flexbox.style.webkitFlexFlow is "column wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column wrap-reverse"
PASS flexbox.style.webkitFlexFlow is "column-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column-reverse none"
PASS flexbox.style.webkitFlexFlow is "column-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column-reverse none"
PASS flexbox.style.webkitFlexFlow is "column-reverse none"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column-reverse none"
PASS flexbox.style.webkitFlexFlow is "column-reverse none"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column-reverse none"
PASS flexbox.style.webkitFlexFlow is "column-reverse wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column-reverse wrap"
PASS flexbox.style.webkitFlexFlow is "column-reverse wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column-reverse wrap"
PASS flexbox.style.webkitFlexFlow is "column-reverse wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column-reverse wrap-reverse"
PASS flexbox.style.webkitFlexFlow is "column-reverse wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column-reverse wrap-reverse"
PASS flexbox.style.webkitFlexFlow is ""
PASS flexbox.style.webkitFlexFlow is ""
PASS flexbox.style.webkitFlexFlow is ""
PASS flexbox.style.webkitFlexFlow is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row none"
PASS flexbox.style.webkitFlexFlow is "column"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column none"
PASS flexbox.style.webkitFlexFlow is "column wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column wrap"
PASS flexbox.style.webkitFlexFlow is "row-reverse wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row-reverse wrap-reverse"
PASS flexbox.style.webkitFlexLinePack is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexLinePack is "stretch"
PASS flexbox.style.webkitFlexLinePack is "start"
PASS window.getComputedStyle(flexbox, null).webkitFlexLinePack is "start"
PASS flexbox.style.webkitFlexLinePack is "end"
PASS window.getComputedStyle(flexbox, null).webkitFlexLinePack is "end"
PASS flexbox.style.webkitFlexLinePack is "center"
PASS window.getComputedStyle(flexbox, null).webkitFlexLinePack is "center"
PASS flexbox.style.webkitFlexLinePack is "space-between"
PASS window.getComputedStyle(flexbox, null).webkitFlexLinePack is "space-between"
PASS flexbox.style.webkitFlexLinePack is "space-around"
PASS window.getComputedStyle(flexbox, null).webkitFlexLinePack is "space-around"
PASS flexbox.style.webkitFlexLinePack is "stretch"
PASS window.getComputedStyle(flexbox, null).webkitFlexLinePack is "stretch"
PASS flexbox.style.webkitFlexLinePack is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexLinePack is "stretch"
PASS flexbox.style.webkitFlexLinePack is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexLinePack is "stretch"
PASS successfullyParsed is true

TEST COMPLETE
